Refactor/lab03/cognito

+++ LAB03/03-Cognito/backend/cloudalbum/database/__init__.py
@@ -1,15 +1,33 @@
-from flask import current_app as app
-from cloudalbum.database.model_ddb import Photo
+"""
+    cloudalbum/database/__init__.py
+    ~~~~~~~~~~~~~~~~~~~~~~~
+    Functions for DB table handling
 
-if not Photo.exists():
-    Photo.create_table(read_capacity_units=app.config['DDB_RCU'], write_capacity_units=app.config['DDB_WCU'], wait=True)
-    print('DynamoDB Photo table created!')
+    :description: CloudAlbum is a fully featured sample application for 'Moving to AWS serverless' training course
+    :copyright: © 2019 written by Dayoungle Jun, Sungshik Jou.
+    :license: MIT, see LICENSE for more details.
+"""
+from cloudalbum.database.model_ddb import Photo
+from flask import current_app as app
+
 
 def create_table():
-    Photo.create_table()
-    print("Dynamodb Photos table created")
-    pass
+    ### User table is no longer needed
+    # if not User.exists():
+    #     app.logger.debug('Creating DynamoDB User table..')
+    #     User.create_table(read_capacity_units=app.config['DDB_RCU'],
+    #                       write_capacity_units=app.config['DDB_WCU'],
+    #                       wait=True)
+    if not Photo.exists():
+        app.logger.debug('Creating DynamoDB Photo table..')
+        Photo.create_table(read_capacity_units=app.config['DDB_RCU'],
+                           write_capacity_units=app.config['DDB_WCU'],
+                           wait=True)
+
 
 def delete_table():
-    Photo.delete_table()
-    print("Dynamodb Users table deleted")
+    ### User table is no longer needed
+    # if User.exists():
+    #     User.delete_table()
+    if Photo.exists():
+        Photo.delete_table()
